Move sight thumbnailing functions to own module
